검색 기능 ã추가

+++ src/main/java/com/example/fiction_place1/domain/board/service/BoardService.java
@@ -12,6 +12,7 @@ import com.example.fiction_place1.domain.user.entity.SiteUser;
 import com.example.fiction_place1.domain.user.entity.User;
 import com.example.fiction_place1.domain.user.repository.CompanyUserRepository;
 import com.example.fiction_place1.domain.user.repository.SiteUserRepository;
+import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -29,19 +30,19 @@ public class BoardService {
     // 게시판 타입에 맞는 게시글 조회
     public Page<Board> getBoardType(Long boardTypeId, Pageable pageable) {
         BoardType boardType = boardTypeRepository.findById(boardTypeId)
-                .orElseThrow(() -> new RuntimeException("BoardType not found"));
+                .orElseThrow(() -> new EntityNotFoundException("BoardType not found"));
         return boardRepository.findByBoardTypeOrderByCreatedDateDesc(boardType, pageable);
     }
 
     public void createFreeBoard(String title, String content, Long boardTypeId, User user){
-
         BoardType boardType = boardTypeRepository.findById(boardTypeId)
-                .orElseThrow(() -> new RuntimeException("BoardType not found"));
+                .orElseThrow(() -> new EntityNotFoundException("BoardType not found"));
 
         Board board = new Board();
         board.setTitle(title);
         board.setContent(content);
         board.setBoardType(boardType);
+
         // 게시글 작성자 설정
         if (user instanceof SiteUser) {
             SiteUser siteUser = (SiteUser) user;
@@ -52,19 +53,31 @@ public class BoardService {
         }
         this.boardRepository.save(board);
     }
+
     public Board getBoard(Long id){
-        Board board = boardRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Board not found"));
-        return board;
+        return boardRepository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException("Board not found"));
     }
+
     public void modify(Board board, String title, String content){
         board.setTitle(title);
         board.setContent(content);
         this.boardRepository.save(board);
     }
+
     public void delete(Board board){
-        this.boardRepository.delete(board);
+        if (boardRepository.existsById(board.getId())) {
+            this.boardRepository.delete(board);
+        } else {
+            throw new EntityNotFoundException("Board not found for deletion");
+        }
     }
+    // 게시글 검색 기능
+    public List<Board> searchBoards(String keyword) {
+        return boardRepository.findByTitleContainingOrContentContainingOrCompanyUser_CompanyNameContainingOrSiteUser_NicknameContaining(
+                keyword, keyword, keyword, keyword);
+    }
+
 }
